moved insertMockData in to HibernateUtil
