+++ .idea/workspace.xml
@@ -4,17 +4,11 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="da6eeb10-8cbf-4d71-bdfe-1669fb6230d9" name="Changes" comment="mock data included">
-      <change afterPath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/utils/MockDataUtil.java" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources.local.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/dataSources.local.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/dataSources.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources/616fb8df-7b0e-4544-ada5-04cd62b84605/storage_v2/_src_/schema/information_schema.FNRwLQ.meta" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources/616fb8df-7b0e-4544-ada5-04cd62b84605/storage_v2/_src_/schema/mysql.osA4Bg.meta" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources/616fb8df-7b0e-4544-ada5-04cd62b84605/storage_v2/_src_/schema/performance_schema.kIw0nw.meta" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources/616fb8df-7b0e-4544-ada5-04cd62b84605/storage_v2/_src_/schema/sys.zb4BAA.meta" beforeDir="false" />
-      <change beforePath="$PROJECT_DIR$/.idea/dataSources/616fb8df-7b0e-4544-ada5-04cd62b84605/storage_v2/_src_/schema/wigellsconcert.g6M7Jw.meta" beforeDir="false" />
+    <list default="true" id="da6eeb10-8cbf-4d71-bdfe-1669fb6230d9" name="Changes" comment="updated mockdata to update immidietely program starta">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/utils/HibernateUtil.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/utils/HibernateUtil.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/utils/MockDataUtil.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/java/com/example/wiggelsconcert/utils/MockDataUtil.java" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -165,7 +159,7 @@
       <workItem from="1741861482603" duration="2190000" />
       <workItem from="1741863723269" duration="539000" />
       <workItem from="1741864302690" duration="844000" />
-      <workItem from="1741865196387" duration="341000" />
+      <workItem from="1741865196387" duration="3559000" />
     </task>
     <task id="LOCAL−00001" summary="uppdatera">
       <option name="closed" value="true" />
@@ -223,7 +217,14 @@
       <option name="presentableId" value="LOCAL−00008" />
       <updated>1741858831990</updated>
     </task>
-    <option name="localTasksCounter" value="9" />
+    <task id="LOCAL−00009" summary="updated mockdata to update immidietely program starta">
+      <option name="closed" value="true" />
+      <created>1741865631719</created>
+      <option name="number" value="LOCAL−00009" />
+      <option name="presentableId" value="LOCAL−00009" />
+      <updated>1741865631719</updated>
+    </task>
+    <option name="localTasksCounter" value="10" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -237,6 +238,7 @@
     <MESSAGE value="doas creation completed" />
     <MESSAGE value="logiken för att ta fram alla kunder som köpt biljett till specifik konsert" />
     <MESSAGE value="mock data included" />
-    <option name="LAST_COMMIT_MESSAGE" value="mock data included" />
+    <MESSAGE value="updated mockdata to update immidietely program starta" />
+    <option name="LAST_COMMIT_MESSAGE" value="updated mockdata to update immidietely program starta" />
   </component>
 </project>
