+++ src/main/java/com/example/wiggelsconcert/utils/HibernateUtil.java
@@ -1,6 +1,8 @@
 package com.example.wiggelsconcert.utils;
 
+import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
 import org.hibernate.cfg.Configuration;
 
 public class HibernateUtil {
@@ -25,6 +27,7 @@ public class HibernateUtil {
             configuration.setProperty("hibernate.connection.password", password);
 
             sessionFactory = configuration.buildSessionFactory();
+            insertMockData();
             sessionFactory.openSession().close();
 
             return true;
@@ -43,6 +46,25 @@ public class HibernateUtil {
         if (sessionFactory != null) {
             sessionFactory.close();
             sessionFactory = null;
+        }
+    }
+
+    private static void insertMockData() {
+        SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
+        Session session = sessionFactory.openSession();
+        Transaction transaction = session.beginTransaction();
+
+        try {
+            MockDataUtil.insertMockData(session);
+            transaction.commit();
+            System.out.println("✅ Mock data inserted successfully!");
+        } catch (Exception e) {
+            if (transaction != null) {
+                transaction.rollback();
+            }
+            System.err.println("Error inserting mock data: " + e.getMessage());
+        } finally {
+            session.close();
         }
     }
 }
